+++ tests/unit/console_results_test.sh
@@ -26,7 +26,7 @@ function test_not_render_passed_assertions_when_no_passed_tests_nor_assertions()
 function test_render_passed_tests_when_passed_tests() {
   local TESTS_PASSED=1
 
-  assertMatches\
+  assert_matches\
     $'Tests:[^\n]*\e\[32m1 passed\e\[0m[^\n]*total'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -35,7 +35,7 @@ function test_render_passed_tests_when_passed_assertions() {
   local TESTS_PASSED=0
   local ASSERTIONS_PASSED=1
 
-  assertMatches\
+  assert_matches\
     $'Tests:[^\n]*\e\[32m0 passed\e\[0m[^\n]*total'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -44,7 +44,7 @@ function test_render_passed_assertions_when_passed_tests() {
   local TESTS_PASSED=1
   local ASSERTIONS_PASSED=0
 
-  assertMatches\
+  assert_matches\
     $'Assertions:[^\n]*\e\[32m0 passed\e\[0m[^\n]*total'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -52,7 +52,7 @@ function test_render_passed_assertions_when_passed_tests() {
 function test_render_passed_assertions_when_passed_assertions() {
   local ASSERTIONS_PASSED=1
 
-  assertMatches\
+  assert_matches\
     $'Assertions:[^\n]*\e\[32m1 passed\e\[0m[^\n]*total'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -76,7 +76,7 @@ function test_not_render_failed_assertions_when_not_failed_tests() {
 function test_render_failed_tests_when_failed_tests() {
   local TESTS_FAILED=1
 
-  assertMatches\
+  assert_matches\
     $'Tests:[^\n]*\e\[31m1 failed\e\[0m[^\n]*total'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -85,7 +85,7 @@ function test_render_failed_assertions_when_failed_tests() {
   local TESTS_FAILED=1
   local ASSERTIONS_FAILED=0
 
-  assertMatches\
+  assert_matches\
     $'Assertions:[^\n]*\e\[31m0 failed\e\[0m[^\n]*total'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -101,7 +101,7 @@ function test_not_render_all_tests_passed_when_failed_tests() {
 function test_render_all_tests_passed_when_not_failed_tests() {
   local TESTS_FAILED=0
 
-  assertMatches\
+  assert_matches\
     $'\e\[42mAll tests passed\e\[0m'\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -110,7 +110,7 @@ function test_total_tests_is_the_sum_of_passed_and_failed_tests() {
   local TESTS_PASSED=4
   local TESTS_FAILED=2
 
-  assertMatches\
+  assert_matches\
     "Tests:[^\n]*6 total"\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
@@ -119,14 +119,14 @@ function test_total_asserts_is_the_sum_of_passed_and_failed_asserts() {
   local ASSERTIONS_PASSED=1
   local ASSERTIONS_FAILED=3
 
-  assertMatches\
+  assert_matches\
     "Assertions:[^\n]*4 total"\
     "$(Console::renderResult $TESTS_PASSED $TESTS_FAILED $ASSERTIONS_PASSED $ASSERTIONS_FAILED)"
 }
 
 function test_render_time_of_execution_when_all_assertions_passed() {
   if [[ $_OS != "OSX" ]]; then
-    assertMatches\
+    assert_matches\
       "Time taken: [[:digit:]]+ ms"\
       "$(Console::renderResult)"
   fi
@@ -134,7 +134,7 @@ function test_render_time_of_execution_when_all_assertions_passed() {
 
 function test_render_time_of_execution_when_not_all_assertions_passed() {
   if [[ $_OS != "OSX" ]]; then
-    assertMatches\
+    assert_matches\
       "Time taken: [[:digit:]]+ ms"\
       "$(Console::renderResult)"
   fi