+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - `assertNotEmpty` -> `assert_not_empty`
   - `assertContains` -> `assert_contains`
   - `assertNotContains` -> `assert_not_contains`
+  - `assertMatches` -> `assert_matches`
 
 ## [0.6.0](https://github.com/TypedDevs/bashunit/compare/0.5.0...0.6.0) - 2023-09-19